# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Instagram Clone API",
        "description": "API do projeto pessoal - clone do Instagram feito em Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Feed de posts",
                "operationId": "5f292ed47bdbfa79356750be0807450c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de posts com usuário, imagens e contagens"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Criar um post com até 10 imagens",
                "operationId": "c2e0b0e26215e6d34ece33ebae25f7df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "images"
                                ],
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "example": "Meu novo post"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        },
                                        "maxItems": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/archived": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Listar posts arquivados do próprio usuário",
                "operationId": "da793d285c750da26393588ce75465eb",
                "responses": {
                    "200": {
                        "description": "Lista de posts arquivados"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{post}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Ver detalhes de um post",
                "operationId": "19d7b55d7d5efd7bc27ed7c01d74ca95",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do post"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Editar legenda de um post",
                "operationId": "97ae61bfc5c99543e90d6f4136332c12",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "caption": {
                                        "type": "string",
                                        "example": "Legenda atualizada"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post atualizado"
                    },
                    "403": {
                        "description": "Não é o dono do post"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Excluir um post",
                "operationId": "f5991cc3c6e9a5fcb5b97135fbe27b0c",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post excluído"
                    },
                    "403": {
                        "description": "Não é o dono do post"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{post}/images/{imageId}": {
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Remover uma imagem de um post",
                "operationId": "187a6b792cbf0dbdf77ae6740daa0a2a",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "imageId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Imagem removida"
                    },
                    "403": {
                        "description": "Não é o dono do post"
                    },
                    "422": {
                        "description": "Não é possível remover a única imagem"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{post}/archive": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Arquivar um post",
                "operationId": "945a8c810c0ba3e1bbcc2618845f7295",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post arquivado"
                    },
                    "403": {
                        "description": "Não é o dono do post"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/posts/{post}/unarchive": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Desarquivar um post",
                "operationId": "094408eda160c90067f68d9687c35bb3",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post desarquivado"
                    },
                    "403": {
                        "description": "Não é o dono do post"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}